+++ README.md
@@ -1,2 +1,2 @@
-# Jogo-da-Vida
-Controle de versão para um protótipo do Jogo da Vida, jogo de tabuleiro.
+# sourceCode.Jogo-da-Vida
+Controle de versão para um protótipo do sourceCode.Jogo da Vida, jogo de tabuleiro.
