+++ src/sourceCode/AppJogoDaVida.java
@@ -1,3 +1,5 @@
+package sourceCode;
+
 import java.util.Scanner;
 import java.util.Random;
 import java.io.File;
